Accessories almost working. Initializing seats and adding accessories on PushToWorld. Fixed a few of the issues that vehicles caused. They are kind of safe. TODO: Fix some issues that happen when a vehicle has its first passenger. Fix not being able to change seats into a seat with an Accessory that is a Vehicle.

+++ Trunk/src/hearthstone-world/Unit.h
@@ -27,7 +27,7 @@ class AIInterface;
 #define MAX_PASSIVE_AURAS 192   // grep: i mananged to break this.. :p seems we need more
 
 bool SERVER_DECL Rand(float);
-#define MAKE_ACTION_BUTTON(A,T) (uint32(A) | (uint32(T) << 24))
+#define MAKE_ACTION_BUTTON(A,T) uint32(uint32(A) | (uint32(T) << 24))
 #define UF_TARGET_DIED  1
 #define UF_ATTACKING	2 // this unit is attacking it's selection
 #define SPELL_GROUPS	96
